fix: renaming and fixes
