+++ src/App.js
@@ -2,7 +2,6 @@ import React, { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { BrowserRouter } from "react-router-dom";
 import { Helmet } from 'react-helmet';
-import SessionTimeoutDialog from '../Common/ViewSettings/SessionTimeoutDialog';
 
 import ErrorBoundary from './pages/Fallback/ErrorBoundary';
 import Shell from './components/Shell/Shell';
@@ -10,6 +9,7 @@ import Routes from './routes/Routes';
 
 import './App.css';
 import CenteredContent from './components/Layout/CenteredContent';
+import SessionTimeoutDialog from './components/SessionTimeout/SessionTimeoutDialog';
 
 function App() {
   const { t } = useTranslation();
